[SYCL] migrate sycl-module-split to llvm-split
Removes sycl-module-split tool and replaces all usages with llvm-split
tool.
