+++ llvm/test/tools/sycl-post-link/device-code-split/auto-module-split-func-ptr.ll
@@ -4,7 +4,7 @@
 ; RUN: FileCheck %s -input-file=%t_0.ll --check-prefix=CHECK-IR0
 ; RUN: FileCheck %s -input-file=%t_1.ll --check-prefix=CHECK-IR1
 
-; RUN: sycl-module-split -split=auto -S < %s -o %t2
+; RUN: llvm-split -sycl-split=auto -S < %s -o %t2
 ; RUN: FileCheck %s -input-file=%t2_0.sym --check-prefix=CHECK-SYM0
 ; RUN: FileCheck %s -input-file=%t2_1.sym --check-prefix=CHECK-SYM1
 ; RUN: FileCheck %s -input-file=%t2_0.ll --check-prefix=CHECK-IR0
