+++ llvm/test/tools/sycl-post-link/device-code-split/per-aspect-split-3.ll
@@ -17,7 +17,7 @@
 ; RUN: FileCheck %s -input-file=%t_1.ll --check-prefix CHECK-M1-IR \
 ; RUN:     --implicit-check-not kernel0 --implicit-check-not bar
 
-; RUN: sycl-module-split -split=auto -S < %s -o %t2
+; RUN: llvm-split -sycl-split=auto -S < %s -o %t2
 ; RUN: FileCheck %s -input-file=%t2.table --check-prefix CHECK-TABLE
 ;
 ; RUN: FileCheck %s -input-file=%t2_0.sym --check-prefix CHECK-M0-SYMS \
