+++ llvm/test/tools/sycl-post-link/device-code-split/per-reqd-sub-group-size-split-2.ll
@@ -16,7 +16,7 @@
 ; RUN:     --implicit-check-not kernel0 --implicit-check-not kernel1 \
 ; RUN:     --implicit-check-not kernel2
 
-; RUN: sycl-module-split -split=auto -S %s -o %t2
+; RUN: llvm-split -sycl-split=auto -S %s -o %t2
 ; RUN: FileCheck %s -input-file=%t2.table --check-prefix CHECK-TABLE
 ;
 ; RUN: FileCheck %s -input-file=%t2_0.sym --check-prefix CHECK-M0-SYMS \
