+++ llvm/test/tools/sycl-post-link/device-code-split/vtable.ll
@@ -42,7 +42,7 @@
 ; RUN: sycl-post-link -properties -split=auto -S < %s -o %t.table
 ; RUN: FileCheck %s -input-file=%t_0.ll
 ;
-; RUN: sycl-module-split -split=auto -S < %s -o %t2
+; RUN: llvm-split -sycl-split=auto -S < %s -o %t2
 ; RUN: FileCheck %s -input-file=%t2_0.ll
 ;
 ; CHECK-DAG: @_ZTV8Derived1 = {{.*}} @_ZN8Derived17displayEv
